I found out you have to check the closing notification's index at the moment it is closing b/c it might not be at the original `index`
Can't `notifications.splice(index, 1);` here because
``notifications[index]` may not exist at this point due to a previous
`notifications.splice(index, 1);`. We have to get the index at the time
it is actually going to close. Found that some of the old notification
objects were hanging around in `notifications` without this fix.